Check availability before querying backend
This change removes a lot of deprecated code pertaining to the secure
settings controlling the feature availability. Under the new model,
this is handled completely externally from Wallet, and we just read
the current state of the value when invoking Global Actions.

Because we can now query backend availability synchronously, rather
than require the BackendCardManager caller to check if the
QuickAccessWalletClient has an available WalletService before querying
the backend (resulting in a hang), we can perform that check in the
BackendComponent when generating the UI-scope subcomponent; if the
check fails, then the backend is unavailable, and so the subcomponent
cannot be created, and we return null, which results in the "vertical"
power menu being shown (and no Wallet).

Fixes: 148794788
Test: atest, manual
Change-Id: I6f4678d298e2dd24da317a074704dd737f212256